Move SwipeUpDown from vendor to shared
